Verify the checksums of apr and openssl tarballs
Motivation:

These sources are downloaded over plaintext connections. We must verify
the known checksum to ensure the file has not been tampered with.

Modifications:

Utilize the antrun checksum task to verify the expected checksum of apr
and openssl.

Result:

The build will fail in the event of a mismatched checksum.